feat: Add "My Classroom" to dashboard and deep linking for classroom
Introduces a new "My Classroom" section to the user dashboard, providing a dedicated space for users to access their training.

Additionally, implements deep linking for the Classroom portal. This allows users to be directed to a specific tab within the classroom when navigating from external links or state changes, improving user experience and onboarding flow.